Cursor type
Cursors are opaque to the client, they may use them as bookmarks but any
more internal semantics may change without notice. E.g. we might need to
move from indices (the plan now) to MongoDB cursors to ID's to anything
else. That's why we Base64-encode them, to make it less appealing to the
client to start to do anything with them.